style: apply suggested edits
