+++ std/algebra/emulated/sw_bn254/pairing_test.go
@@ -8,9 +8,6 @@ import (
 	"github.com/consensys/gnark-crypto/ecc"
 	"github.com/consensys/gnark-crypto/ecc/bn254"
 	"github.com/consensys/gnark/frontend"
-	"github.com/consensys/gnark/std/algebra/emulated/fields_bn254"
-	"github.com/consensys/gnark/std/algebra/emulated/sw_emulated"
-	"github.com/consensys/gnark/std/math/emulated"
 	"github.com/consensys/gnark/test"
 )
 
@@ -29,9 +26,9 @@ func randomG1G2Affines(assert *test.Assert) (bn254.G1Affine, bn254.G2Affine) {
 }
 
 type MillerLoopCircuit struct {
-	InG1 sw_emulated.AffinePoint[emulated.BN254Fp]
+	InG1 G1Affine
 	InG2 G2Affine
-	Res  fields_bn254.GTEl
+	Res  GTEl
 }
 
 func (c *MillerLoopCircuit) Define(api frontend.API) error {
@@ -43,7 +40,7 @@ func (c *MillerLoopCircuit) Define(api frontend.API) error {
 	if err != nil {
 		return fmt.Errorf("pair: %w", err)
 	}
-	pairing.Ext12.AssertIsEqual(res, &c.Res)
+	pairing.AssertIsEqual(res, &c.Res)
 	return nil
 }
 
@@ -55,15 +52,15 @@ func TestMillerLoopTestSolve(t *testing.T) {
 	witness := MillerLoopCircuit{
 		InG1: NewG1Affine(p),
 		InG2: NewG2Affine(q),
-		Res:  fields_bn254.NewGTEl(res),
+		Res:  NewGTEl(res),
 	}
 	err = test.IsSolved(&MillerLoopCircuit{}, &witness, ecc.BN254.ScalarField())
 	assert.NoError(err)
 }
 
 type FinalExponentiationCircuit struct {
-	InGt fields_bn254.GTEl
-	Res  fields_bn254.GTEl
+	InGt GTEl
+	Res  GTEl
 }
 
 func (c *FinalExponentiationCircuit) Define(api frontend.API) error {
@@ -72,7 +69,7 @@ func (c *FinalExponentiationCircuit) Define(api frontend.API) error {
 		return fmt.Errorf("new pairing: %w", err)
 	}
 	res := pairing.FinalExponentiation(&c.InGt)
-	pairing.Ext12.AssertIsEqual(res, &c.Res)
+	pairing.AssertIsEqual(res, &c.Res)
 	return nil
 }
 
@@ -82,8 +79,8 @@ func TestFinalExponentiationTestSolve(t *testing.T) {
 	gt.SetRandom()
 	res := bn254.FinalExponentiation(&gt)
 	witness := FinalExponentiationCircuit{
-		InGt: fields_bn254.NewGTEl(gt),
-		Res:  fields_bn254.NewGTEl(res),
+		InGt: NewGTEl(gt),
+		Res:  NewGTEl(res),
 	}
 	err := test.IsSolved(&FinalExponentiationCircuit{}, &witness, ecc.BN254.ScalarField())
 	assert.NoError(err)
@@ -92,7 +89,7 @@ func TestFinalExponentiationTestSolve(t *testing.T) {
 type PairCircuit struct {
 	InG1 G1Affine
 	InG2 G2Affine
-	Res  fields_bn254.GTEl
+	Res  GTEl
 }
 
 func (c *PairCircuit) Define(api frontend.API) error {
@@ -104,7 +101,7 @@ func (c *PairCircuit) Define(api frontend.API) error {
 	if err != nil {
 		return fmt.Errorf("pair: %w", err)
 	}
-	pairing.Ext12.AssertIsEqual(res, &c.Res)
+	pairing.AssertIsEqual(res, &c.Res)
 	return nil
 }
 
@@ -116,7 +113,7 @@ func TestPairTestSolve(t *testing.T) {
 	witness := PairCircuit{
 		InG1: NewG1Affine(p),
 		InG2: NewG2Affine(q),
-		Res:  fields_bn254.NewGTEl(res),
+		Res:  NewGTEl(res),
 	}
 	err = test.IsSolved(&PairCircuit{}, &witness, ecc.BN254.ScalarField())
 	assert.NoError(err)
